Refactor widget styles
