+++ forui/lib/src/widgets/card/card_content.dart
@@ -1,7 +1,7 @@
 part of 'card.dart';
 
-/// Represents a content for [FCard].
-final class FCardContent extends StatelessWidget {
+/// Represents a [FCard]'s content.
+@internal final class FCardContent extends StatelessWidget {
   /// The title.
   final String? title;
 
@@ -37,4 +37,58 @@ final class FCardContent extends StatelessWidget {
       ),
     );
   }
+}
+
+/// [FCardContent]'s style.
+final class FCardContentStyle with Diagnosticable {
+  /// The padding.
+  final EdgeInsets padding;
+
+  /// The title.
+  final TextStyle title;
+
+  /// The subtitle.
+  final TextStyle subtitle;
+
+  /// Creates a [FCardContentStyle].
+  const FCardContentStyle({required this.padding, required this.title, required this.subtitle});
+
+  /// Creates a [FCardContentStyle] that inherits its properties from [colorScheme].
+  FCardContentStyle.inherit({required FColorScheme colorScheme}):
+    padding = const EdgeInsets.fromLTRB(16, 12, 16, 16),
+    title = TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.w600,
+      color: colorScheme.foreground,
+    ),
+    subtitle = TextStyle(
+      fontSize: 12,
+      color: colorScheme.mutedForeground,
+    );
+
+  /// Creates a copy of this [FCardContentStyle] with the given properties replaced.
+  FCardContentStyle copyWith({EdgeInsets? padding, TextStyle? title, TextStyle? subtitle}) => FCardContentStyle(
+      padding: padding ?? this.padding,
+      title: title ?? this.title,
+      subtitle: subtitle ?? this.subtitle,
+    );
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty<EdgeInsets>('padding', padding))
+      ..add(DiagnosticsProperty<TextStyle>('title', title))
+      ..add(DiagnosticsProperty<TextStyle>('subtitle', subtitle));
+  }
+
+  @override
+  bool operator ==(Object other) => identical(this, other) || other is FCardContentStyle &&
+      runtimeType == other.runtimeType &&
+      padding == other.padding &&
+      title == other.title &&
+      subtitle == other.subtitle;
+
+  @override
+  int get hashCode => padding.hashCode ^ title.hashCode ^ subtitle.hashCode;
 }
